pyproject.toml
[tool.poetry]
name = "fast-api-proj"
version = "0.1.0"
description = ""
authors = ["nwafor6 <nwaforglory6@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.5"
uvicorn = "^0.32.0"
python-dotenv = "^1.0.1"
boto3 = "^1.35.63"
pynamodb = "^6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
